% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{add_tooltip_icon}
\alias{add_tooltip_icon}
\title{Add a tooltip icon to a label}
\usage{
add_tooltip_icon(..., id, icon = shiny::icon("info-circle"))
}
\arguments{
\item{...}{Arguments passed to span element (e.g., the label to add the
tooltip to).}

\item{id}{The id of the tooltip.}

\item{icon}{The icon to use for the tooltip.}
}
\value{
The label with the tooltip icon.
}
\description{
Add a tooltip icon to a label
}
