% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{checkbox}
\alias{checkbox}
\title{Wrapper around \code{prettyCheckbox()}}
\usage{
checkbox(
  inputId,
  label,
  value = FALSE,
  status = "primary",
  animation = "jelly",
  icon = shiny::icon("check"),
  bigger = TRUE,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{status}{Add a class to the checkbox,
you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.}

\item{animation}{Add an animation when checkbox is checked, a value between
\code{smooth}, \code{jelly}, \code{tada}, \code{rotate}, \code{pulse}.}

\item{icon}{Optional, display an icon on the checkbox, must be an icon created with \code{icon}.}

\item{bigger}{Scale the checkboxes a bit bigger (\code{TRUE} or \code{FALSE}).}

\item{...}{Additional arguments to pass to \code{shinyWidgets::prettyCheckbox()}.}
}
\description{
Wrapper around \code{prettyCheckbox()} with different default
settings.
}
