% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{continue_button}
\alias{continue_button}
\title{Continue button}
\usage{
continue_button(inputId = "continue", label = "Next stage", ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{...}{Additional arguments to pass to \code{shiny::actionButton()}.}
}
\description{
Customized action button using \code{actionButton()}.
}
