% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{css_styler}
\alias{css_styler}
\title{Add css styling}
\usage{
css_styler(...)
}
\arguments{
\item{...}{Any number of named arguments with the name being the css
selector and the value being the css style value.}
}
\description{
Add css styling
}
