% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{download_button}
\alias{download_button}
\title{Download button}
\usage{
download_button(inputId, label = "", style = "simple", css_style = NULL, ...)
}
\arguments{
\item{inputId}{Input ID.}

\item{label}{The contents of the button, usually a text label.}

\item{style}{Style of the button, to choose between \code{simple}, \code{bordered},
\code{minimal}, \code{stretch}, \code{jelly}, \code{gradient}, \code{fill},
\code{material-circle}, \code{material-flat}, \code{pill}, \code{float}, \code{unite}.}

\item{css_style}{CSS style for the button.}

\item{...}{Additional arguments to pass to \code{shinyWidgets::downloadBttn()}.}
}
\description{
Wrapper around \code{downloadBttn()} with different default settings
}
