% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-file-input.R
\name{fileInputModule}
\alias{fileInputModule}
\alias{fileInputUI}
\alias{fileInputServer}
\title{File input module}
\usage{
fileInputUI(
  id,
  label = "File Upload",
  accept = c(".csv", ".txt", ".rds"),
  show_accept = TRUE,
  tooltip = NULL
)

fileInputServer(id, default_data = NULL)
}
\arguments{
\item{id}{Unique identifier.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{accept}{A character vector of "unique file type specifiers" which gives
the browser a hint as to the type of file the server expects. Many browsers
use this prevent the user from selecting an invalid file.

A unique file type specifier can be:
\itemize{
\item A case insensitive extension like \code{.csv} or \code{.rds}.
\item A valid MIME type, like \code{text/plain} or \code{application/pdf}
\item One of \verb{audio/*}, \verb{video/*}, or \verb{image/*} meaning any audio, video,
or image type, respectively.
}}

\item{tooltip}{Tooltip to display.}
}
\description{
Shiny module (UI and server-side functions) to upload a file.
}
\keyword{internal}
