% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltips.R
\name{format_tooltip}
\alias{format_tooltip}
\title{Format tooltips}
\usage{
format_tooltip(tooltip, size = 14)
}
\arguments{
\item{tooltip}{The tooltip to add.}

\item{size}{Font size}
}
\value{
Formatted tooltip
}
\description{
Format tooltips
}
