% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_icon}
\alias{get_icon}
\title{Get html snippet for built-in icons.}
\usage{
get_icon(icon_name, width = 20, height = 20)
}
\arguments{
\item{icon_name}{Name of built-in icon.}

\item{width}{Width of icon in pixels.}

\item{height}{Height of icon in pixels.}
}
\value{
Character vector of icon path.
}
\description{
Get html snippet for built-in icons.
}
