% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_icons}
\alias{get_icons}
\title{Return custom icons}
\usage{
get_icons(icon_names, width = 20, height = 20)
}
\arguments{
\item{icon_names}{Name(s) of built-in icons to return.}

\item{width}{Width of icon in pixels.}

\item{height}{Height of icon in pixels.}
}
\value{
Named character vector of html snippets to display icons
}
\description{
Return custom icons
}
