% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{hr_short}
\alias{hr_short}
\title{Short horizontal rule}
\usage{
hr_short(
  margin_left = "105px",
  margin_right = margin_left,
  color = "lightblue"
)
}
\arguments{
\item{margin_left}{Left margin.}

\item{margin_right}{Right margin.}

\item{color}{Color of line.}
}
\description{
Short horizontal rule
}
