% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui-boxes.R
\name{innerFluidRow}
\alias{innerFluidRow}
\title{Inner fluid row that removes negative margins}
\usage{
innerFluidRow(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::fluidRow()}.}
}
\description{
Wrapper around \code{shiny::fluidRow()} that is useful when used
inside prettyBox().
}
