% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_null_js}
\alias{is_null_js}
\title{Check if input is NULL in conditionalPanel (ie.., Javascript)}
\usage{
is_null_js(inputId, input = TRUE)
}
\arguments{
\item{inputId}{Input id}

\item{input}{If \code{TRUE}, check if \code{input[inputId]} is \code{NULL}.
Otherwise, check if \code{inputId} is \code{NULL}.}
}
\value{
String of Javascript code.
}
\description{
Check if input is NULL in conditionalPanel (ie.., Javascript)
}
