% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{material_switch}
\alias{material_switch}
\title{Wrapper around \code{materialSwitch()}}
\usage{
material_switch(
  inputId,
  label,
  value = TRUE,
  status = "primary",
  right = FALSE,
  inline = FALSE,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Input label.}

\item{value}{TRUE or FALSE.}

\item{status}{Color, must be a valid Bootstrap status : default, primary, info, success, warning, danger.}

\item{right}{Should the the label be on the right? default to FALSE.}

\item{inline}{Display the input inline, if you want to place buttons next to each other.}

\item{...}{Additional arguments to pass to \code{shinyWidgets::materialSwitch()}.}
}
\description{
Wrapper around \code{materialSwitch()} with different default
settings.
}
