% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{picker_multiple_input}
\alias{picker_multiple_input}
\title{Wrapper around \code{pickerInput()} to select multiple choices}
\usage{
picker_multiple_input(
  inputId,
  label,
  choices,
  selected = NULL,
  actionsBox = TRUE,
  maxOptions = NULL,
  options = NULL,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from. If elements of the
list are named then that name rather than the value is displayed to the user.}

\item{selected}{The initially selected value (or multiple values if \code{multiple = TRUE}).
If not specified then defaults to the first value for single-select lists
and no values for multiple select lists.}

\item{actionsBox}{Logical. Whether or not to include actions box.}

\item{maxOptions}{Numeric. Max number of options that can be selected.}

\item{options}{List of options, see \link[shinyWidgets]{pickerOptions} for all available options.
To limit the number of selection possible, see example below.}

\item{...}{Additional arguments to pass to \code{shinyWidgets::pickerInput()}.}
}
\description{
Wrapper around \code{pickerInput()} to select multiple choices but
with different default settings.
}
