% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-options.R
\name{plotGeomOptionsUI}
\alias{plotGeomOptionsUI}
\title{Geom-specific options for plotting}
\usage{
plotGeomOptionsUI(
  id,
  subsample = FALSE,
  text_size = FALSE,
  geom = NULL,
  options_list = NULL
)
}
\arguments{
\item{id}{Unique identifier.}

\item{subsample}{Logical. If \code{TRUE}, include subsample data options.}

\item{text_size}{Logical. If \code{TRUE}, include text size options.}

\item{geom}{Character string or vector of default geom(s) to display.
If \code{NULL}, inferred from \code{input[ns("display_geom")]}.}

\item{options_list}{List of shiny tags of additional options to include.}
}
\value{
List of shiny tags.
}
\description{
Geom-specific options for plotting
}
