% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-plot.R
\name{plotModule}
\alias{plotModule}
\alias{plotUI}
\alias{plotServer}
\title{Basic plot module}
\usage{
plotUI(id, border = FALSE, spinner = FALSE)

plotServer(
  id,
  plot_fun,
  plot_options = TRUE,
  modes = c("ggplot", "plotly"),
  border = FALSE,
  spinner = TRUE,
  error_msg = ""
)
}
\arguments{
\item{id}{Unique identifier.}

\item{border}{Add border to plot.}

\item{spinner}{Add spinner to plot.}

\item{plot_fun}{(Reactive) function to generate plot.}

\item{plot_options}{Logical. Whether to update plot using default plotting
options set using \code{plotOptionsUI(id)}. Only set to \code{TRUE} if
\code{plotOptionsUI(id)} is called in the UI.}

\item{modes}{Character vector. Plot modes to display. Options are
\code{"ggplot"} and \code{"plotly"}.}

\item{error_msg}{Error message to display if specified mode is not
supported.}
}
\description{
Shiny module (UI and server-side functions) to display plot (
ggplot2 or plotly) in a uniform container.
}
\keyword{internal}
