% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-options.R
\name{plotOptionsUI}
\alias{plotOptionsUI}
\title{Basic Plot Options}
\usage{
plotOptionsUI(
  id,
  multicol = FALSE,
  heatmap = FALSE,
  strip_options = FALSE,
  total_width = 12,
  height = 500,
  x_axis_text_size = 12,
  y_axis_text_size = 12,
  legend_text_size = 12,
  strip_text_size = 14,
  x_axis_title_size = 14,
  y_axis_title_size = 14,
  legend_title_size = 14,
  title_size = 16,
  axis_line_width = 1,
  x_text_angle = FALSE,
  strip_text_color = "white"
)
}
\arguments{
\item{id}{Unique identifier.}

\item{multicol}{Logical. If \code{TRUE}, use multiple columns.}

\item{heatmap}{Logical. If \code{TRUE}, include heatmap options.}

\item{strip_options}{Logical. If \code{TRUE}, include strip options.}

\item{total_width}{Total width of html element.}

\item{height}{Height of html element.}

\item{x_axis_text_size}{Default size of x-axis text.}

\item{y_axis_text_size}{Default size of y-axis text.}

\item{legend_text_size}{Default size of legend text.}

\item{strip_text_size}{Default size of strip text.}

\item{x_axis_title_size}{Default size of x-axis title.}

\item{y_axis_title_size}{Default size of y-axis title.}

\item{legend_title_size}{Default size of legend title.}

\item{title_size}{Default size of title.}

\item{axis_line_width}{Default width of axis lines.}

\item{x_text_angle}{Default angle of x-axis text. If \code{FALSE}, x-axis
text is horizontal. Otherwise, it is angled 45 degrees.}

\item{strip_text_color}{Default color of strip text.}
}
\value{
List of shiny tags.
}
\description{
Basic Plot Options
}
