% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{prettyDashboardHeader}
\alias{prettyDashboardHeader}
\title{Pretty dashboard header}
\usage{
prettyDashboardHeader(
  title,
  title_position = c("left", "center"),
  title_style = list(),
  ...
)
}
\arguments{
\item{title}{An optional title to show in the header bar.. By default, this
will also be used as the title shown in the browser's title bar. If you
want that to be different from the text in the dashboard header bar, set
the \code{title} in \code{\link[shinydashboard]{dashboardPage}}.}

\item{title_position}{Position of the title, either "left" or "center".}

\item{title_style}{(Optional) list of CSS styling options for the title.}

\item{...}{Arguments passed to \code{shinydashboard::dashboardHeader()}.}
}
\description{
Wrapper around \code{shinydashboard::dashboardHeader()} that allows
for the title to be positioned to the left or center.
}
