% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{radio_buttons}
\alias{radio_buttons}
\title{Wrapper around \code{prettyRadioButtons()}}
\usage{
radio_buttons(
  inputId,
  label,
  choices,
  selected = NULL,
  status = "primary",
  animation = "jelly",
  icon = shiny::icon("check"),
  bigger = TRUE,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control.}

\item{choices}{List of values to show radio buttons for. If elements of the list
are named then that name rather than the value is displayed to the user. If
this argument is provided, then \code{choiceNames} and \code{choiceValues} must not be provided,
and vice-versa. The values should be strings; other types (such as logicals and
numbers) will be coerced to strings.}

\item{selected}{The values that should be initially selected,
(if not specified then defaults to the first value).}

\item{status}{Add a class to the radio,
you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.}

\item{animation}{Add an animation when radio is checked, a value between
\code{smooth}, \code{jelly}, \code{tada}, \code{rotate}, \code{pulse}.}

\item{icon}{Optional, display an icon on the radio, must be an icon created with \code{icon}.}

\item{bigger}{Scale the radio a bit bigger (\code{TRUE} or \code{FALSE}).}

\item{...}{Additional arguments to pass to
\code{shinyWidgets::prettyRadioButtons()}.}
}
\description{
Wrapper around \code{prettyRadioButtons()} but with different
default settings.
}
