% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{radio_group_buttons}
\alias{radio_group_buttons}
\title{Wrapper around \code{radioGroupButtons()}}
\usage{
radio_group_buttons(
  inputId,
  label,
  choices,
  selected = NULL,
  status = "default",
  size = "sm",
  justified = FALSE,
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to select from (if elements of the list are
named then that name rather than the value is displayed to the user). If
this argument is provided, then \code{choiceNames} and \code{choiceValues} must not
be provided, and vice-versa. The values should be strings; other types
(such as logicals and numbers) will be coerced to strings.}

\item{selected}{The initially selected value. If not specified, then it
defaults to the first item in \code{choices}. To start with no items selected,
use \code{character(0)}.}

\item{status}{Add a class to the buttons, you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.
Or use an arbitrary strings to add a custom class, e.g. : with \code{status = "custom-class"}, buttons will have class \code{btn-custom-class}.}

\item{size}{Size of the buttons ('xs', 'sm', 'normal', 'lg')}

\item{justified}{If TRUE, fill the width of the parent div}

\item{...}{Additional arguments to pass to
\code{shinyWidgets::radioGroupButtons()}.}
}
\description{
Wrapper around \code{radioGroupButtons()} but with different
default settings.
}
