% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{reset_inputs}
\alias{reset_inputs}
\title{Reset all inputs}
\usage{
reset_inputs(input, session)
}
\arguments{
\item{input}{Shiny input.}

\item{session}{Shiny session.}
}
\description{
Reset all inputs to their default values. Note that
\code{shinyjs::useShinyjs()} needs to be includes in the UI for this to work.
}
