% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{set_margins}
\alias{set_margins}
\title{Set margins}
\usage{
set_margins(obj, top = NULL, bottom = NULL, left = NULL, right = NULL)
}
\arguments{
\item{obj}{Object to set margins.}

\item{top}{Top margin.}

\item{bottom}{Bottom margin.}

\item{left}{Left margin.}

\item{right}{Right margin.}
}
\description{
Set margins
}
