% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{small_vspace}
\alias{small_vspace}
\title{Add small amount of vertical space}
\usage{
small_vspace(size = "3px")
}
\arguments{
\item{size}{Size of vertical space.}
}
\description{
Add small amount of vertical space
}
