% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-table.R
\name{tableModule}
\alias{tableModule}
\alias{tableUI}
\alias{tableServer}
\title{Basic table module}
\usage{
tableUI(id, border = FALSE, spinner = FALSE)

tableServer(
  id,
  table_fun,
  table_options = TRUE,
  mode = c("DT", "kable"),
  caption = NULL,
  spinner = TRUE,
  ...
)
}
\arguments{
\item{id}{Unique identifier.}
}
\description{
Shiny module (UI and server-side functions) to display table
(kable or DT) in a uniform container.
}
\keyword{internal}
