% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-options.R
\name{tableOptionsUI}
\alias{tableOptionsUI}
\title{Basic Table Options}
\usage{
tableOptionsUI(id, digits = NA, sigfig = FALSE)
}
\arguments{
\item{id}{Unique identifier.}

\item{digits}{Default number of digits to display.}

\item{sigfig}{Logical. If \code{TRUE}, use significant digits.}
}
\value{
List of shiny tags.
}
\description{
Basic Table Options
}
