% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{add_spinner}
\alias{add_spinner}
\title{Add spinner when loading}
\usage{
add_spinner(obj, color = "#18bc9c")
}
\arguments{
\item{obj}{Object to add spinner to.}

\item{color}{The color of the spinner in hex format. Ignored if \code{image} is used.}
}
\description{
Wrapper around \code{shinycssloaders::withSpinner()} with a different
default color.
}
\examples{
## DON'T RUN
# shiny::htmlOutput({NAME OF UI ELEMENT}) \%>\% add_spinner()

}
