% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_icon}
\alias{get_icon}
\title{Get html snippet for built-in icons.}
\usage{
get_icon(src, width = 20, height = 20, ...)
}
\arguments{
\item{src}{Name of built-in icon.}

\item{width}{Width of icon in pixels.}

\item{height}{Height of icon in pixels.}

\item{...}{Additional arguments to pass to \code{htmltools::img()}.}
}
\value{
Character vector of icon path.
}
\description{
Get html snippet for built-in icons.
}
