% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{reset_inputs}
\alias{reset_inputs}
\title{Reset all inputs}
\usage{
reset_inputs(input, session, reset_input_ids = NULL)
}
\arguments{
\item{input}{Shiny input.}

\item{session}{Shiny session.}

\item{reset_input_ids}{A character vector of input IDs to reset. If NULL,
all inputs will be reset.}
}
\description{
Reset all inputs to their default values. Note that
\code{shinyjs::useShinyjs()} needs to be includes in the UI for this to work.
}
