% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-table.R
\name{tableModule}
\alias{tableModule}
\alias{tableUI}
\alias{tableServer}
\title{Basic table module}
\usage{
tableUI(id, border = FALSE, spinner = FALSE)

tableServer(
  id,
  table_fun,
  table_options = TRUE,
  mode = c("DT", "kable"),
  caption = NULL,
  border = FALSE,
  spinner = TRUE,
  error_msg = "",
  ...
)
}
\arguments{
\item{id}{Unique identifier.}

\item{border}{Add border to plot.}

\item{spinner}{Add spinner to plot.}

\item{table_fun}{(Reactive) function to generate table.}

\item{table_options}{Logical. Whether to update table using default table
options set using \code{tableOptionsUI(id)}. Only set to \code{TRUE} if
\code{tableOptionsUI(id)} is called in the UI.}

\item{caption}{Table caption.}

\item{error_msg}{Error message to display if specified mode is not
supported.}

\item{...}{Arguments to pass to \code{vthemes::pretty_DT()} or
\code{vthemes::pretty_kable()}}

\item{modes}{Character string Table mode to display. Options are
\code{"DT"} and \code{"kable"}.}
}
\description{
Shiny module (UI and server-side functions) to display table
(kable or DT) in a uniform container.
}
\keyword{internal}
