% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-text.R
\name{textBoxModule}
\alias{textBoxModule}
\alias{textboxUI}
\alias{textboxServer}
\title{Basic text box module}
\usage{
textboxUI(id)

textboxServer(id, text)
}
\arguments{
\item{id}{Unique identifier.}
}
\description{
Shiny module (UI and server-side functions) to display text.
}
\keyword{internal}
